@@ -1,114 +1,61 @@
 {
   "name": "hippymood",
-  "version": "0.5.0",
-  "description": " Web app to play your music smoothly by genre/mood. ",
-  "author": "Ronan Le Pivaingt <r.lepivaingt@gmail.com>",
+  "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "node build/dev-server.js",
-    "prod": "node build/prod-server.js",
-    "build": "node build/build.js",
-    "dbInit": "node server/controllers/DatabaseController.js Up",
-    "unit": "cross-env BABEL_ENV=test karma start test/unit/karma.conf.js --single-run",
-    "e2e": "node test/e2e/runner.js",
-    "test": "npm run unit && npm run e2e",
-    "lint": "eslint --ext .js,.vue src test/unit/specs test/e2e/specs"
+    "serve": "vue-cli-service serve",
+    "build": "vue-cli-service build",
+    "lint": "vue-cli-service lint"
   },
   "dependencies": {
+    "axios": "^0.18.0",
+    "body-parser": "^1.18.3",
     "connect-session-knex": "^1.4.0",
     "cors": "^2.8.4",
-    "express-session": "^1.15.1",
-    "hchs-vue-charts": "^1.2.8",
+    "es6-promise": "^4.2.5",
+    "express": "^4.16.3",
+    "express-session": "^1.15.6",
     "id3js": "^1.1.3",
-    "intro.js": "^2.9.3",
-    "knex": "^0.13.0",
+    "knex": "^0.15.2",
     "make-runnable": "^1.3.6",
-    "multer": "^1.3.0",
+    "multer": "^1.3.1",
     "musicmetadata": "^2.0.5",
     "mysql": "^2.16.0",
-    "socket.io": "^2.0.3",
-    "v-tooltip": "^2.0.0-rc.33",
-    "vue": "^2.2.2",
-    "vue-clip": "^1.0.0",
-    "vue-introjs": "^1.3.0",
-    "vue-material": "^0.7.1",
-    "vue-progressbar": "^0.7.3",
-    "vue-resource": "^1.2.1",
-    "vue-router": "^2.2.0",
-    "vue-socket.io": "^2.1.1-a",
-    "vue-touch": "^2.0.0-beta.4",
-    "vue-trianglify": "git+https://github.com/multi-cell/vue-trianglify.git#dev",
-    "vue-youtube-embed": "^2.1.1",
-    "vuex": "^2.2.1",
-    "youtube-url": "^0.5.0"
+    "vue": "^2.5.17",
+    "vue-axios": "^2.1.3",
+    "vue-material": "^1.0.0-beta-10.2",
+    "vue-progressbar": "^0.7.5",
+    "vue-router": "^3.0.1",
+    "vuex": "^3.0.1"
   },
   "devDependencies": {
-    "autoprefixer": "^6.7.2",
-    "babel-core": "^6.22.1",
-    "babel-eslint": "^7.1.1",
-    "babel-loader": "^6.2.10",
-    "babel-plugin-istanbul": "^3.1.2",
-    "babel-plugin-transform-runtime": "^6.22.0",
-    "babel-preset-env": "^1.2.1",
-    "babel-preset-stage-2": "^6.22.0",
-    "babel-register": "^6.22.0",
-    "chai": "^3.5.0",
-    "chalk": "^1.1.3",
-    "chromedriver": "^2.27.2",
-    "connect-history-api-fallback": "^1.3.0",
-    "copy-webpack-plugin": "^4.0.1",
-    "cross-env": "^3.1.4",
-    "cross-spawn": "^5.0.1",
-    "css-loader": "^0.26.1",
-    "eslint": "^3.14.1",
-    "eslint-config-standard": "^6.2.1",
-    "eslint-friendly-formatter": "^2.0.7",
-    "eslint-loader": "^1.6.1",
-    "eslint-plugin-html": "^2.0.0",
-    "eslint-plugin-promise": "^3.4.0",
-    "eslint-plugin-standard": "^2.0.1",
-    "eventsource-polyfill": "^0.9.6",
-    "express": "^4.14.1",
-    "extract-text-webpack-plugin": "^2.0.0",
-    "file-loader": "^0.10.0",
-    "friendly-errors-webpack-plugin": "^1.1.3",
-    "function-bind": "^1.1.0",
-    "html-webpack-plugin": "^2.28.0",
-    "http-proxy-middleware": "^0.17.3",
-    "inject-loader": "^2.0.1",
-    "karma": "^1.4.1",
-    "karma-coverage": "^1.1.1",
-    "karma-mocha": "^1.3.0",
-    "karma-phantomjs-launcher": "^1.0.2",
-    "karma-sinon-chai": "^1.2.4",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-spec-reporter": "0.0.26",
-    "karma-webpack": "^2.0.2",
-    "lolex": "^1.5.2",
-    "mocha": "^3.2.0",
-    "nightwatch": "^0.9.12",
-    "opn": "^4.0.2",
-    "optimize-css-assets-webpack-plugin": "^1.3.0",
-    "ora": "^1.1.0",
-    "phantomjs-prebuilt": "^2.1.14",
-    "rimraf": "^2.6.0",
-    "selenium-server": "^3.0.1",
-    "semver": "^5.3.0",
-    "sinon": "^1.17.7",
-    "sinon-chai": "^2.8.0",
-    "url-loader": "^0.5.7",
-    "vue-loader": "^11.1.4",
-    "vue-style-loader": "^2.0.0",
-    "vue-template-compiler": "^2.2.4",
-    "webpack": "^2.2.1",
-    "webpack-bundle-analyzer": "^2.2.1",
-    "webpack-dev-middleware": "^1.10.0",
-    "webpack-hot-middleware": "^2.16.1",
-    "webpack-merge": "^2.6.1"
+    "@vue/cli-plugin-babel": "^3.0.3",
+    "@vue/cli-plugin-eslint": "^3.0.3",
+    "@vue/cli-service": "^3.0.3",
+    "@vue/eslint-config-standard": "^3.0.3",
+    "node-sass": "^4.9.3",
+    "sass-loader": "^7.1.0",
+    "vue-template-compiler": "^2.5.17"
   },
-  "engines": {
-    "node": ">= 4.0.0",
-    "npm": ">= 3.0.0"
+  "eslintConfig": {
+    "root": true,
+    "env": {
+      "node": true
+    },
+    "extends": [
+      "plugin:vue/essential",
+      "eslint:recommended",
+      "@vue/standard"
+    ],
+    "rules": {},
+    "parserOptions": {
+      "parser": "babel-eslint"
+    }
+  },
+  "postcss": {
+    "plugins": {
+      "autoprefixer": {}
+    }
   },
   "browserslist": [
     "> 1%",
