@@ -1,5 +1,5 @@
 // Loading configuration file with database credentials
-var config = require('../config/server.config');
+var config = require('../config');
 var dbConfig = {
   client: 'mysql',
   connection: {
